refactor(studio): move upload helpers into utils.ts
const.ts held constants, a schema, and eight functions. The package uses
utils.ts for this — 34 of them against 2 const.ts, and both of those are
this modal and the CloneAgentModal it was copied from.

Constants and the form schema stay in const.ts; the pure helpers and
AgentConfigParseError move to utils.ts, with const.test.ts renamed to
match. No behaviour change.
