+++ web/packages/studio/src/routes/agents/AgentsListRoute/UploadAgentModal/utils.ts
@@ -1,0 +1,123 @@
+// SPDX-FileCopyrightText: Copyright (c) 2025-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+import {
+  AGENT_CONFIG_FILENAME,
+  AGENT_SPEC_FILENAME,
+  FABRIC_CONFIG_FORMAT,
+  IGNORED_DIRECTORIES,
+  IGNORED_EXTENSIONS,
+  IGNORED_FILENAMES,
+  MAX_AGENT_SPEC_BYTES,
+  MAX_AGENT_SPEC_FILES,
+  MAX_PICKED_FILES,
+} from '@studio/routes/agents/AgentsListRoute/UploadAgentModal/const';
+import type { UploadAgentEntry } from '@studio/routes/agents/AgentsListRoute/UploadAgentModal/type';
+import YAML from 'yaml';
+
+/** Convention only — the Agent entity stores no reference to it. */
+export const agentSpecFilesetName = (agentName: string): string => `${agentName}-spec`;
+
+export const tooManyPickedFiles = (pickedCount: number): string | undefined =>
+  pickedCount > MAX_PICKED_FILES
+    ? `That directory holds ${pickedCount.toLocaleString()} files, far more than an agent directory should. Point at the agent's own directory.`
+    : undefined;
+
+export const isIgnoredPath = (path: string): boolean => {
+  const segments = path.split('/');
+  if (segments.some((segment) => IGNORED_DIRECTORIES.has(segment))) return true;
+
+  const filename = segments[segments.length - 1] ?? '';
+  if (IGNORED_FILENAMES.has(filename)) return true;
+
+  return IGNORED_EXTENSIONS.some((extension) => filename.endsWith(extension));
+};
+
+const pathCollator = new Intl.Collator();
+
+/** The fileset holds the directory's contents, so the picked root is stripped from each path. */
+export const collectAgentEntries = (files: File[]): UploadAgentEntry[] => {
+  const entries: UploadAgentEntry[] = [];
+
+  for (const file of files) {
+    const relativePath = file.webkitRelativePath || file.name;
+    const path = relativePath.split('/').slice(1).join('/') || file.name;
+    if (!path || isIgnoredPath(path)) continue;
+    entries.push({ path, file });
+  }
+
+  return entries.sort((left, right) => pathCollator.compare(left.path, right.path));
+};
+
+export const totalEntryBytes = (entries: UploadAgentEntry[]): number =>
+  entries.reduce((total, entry) => total + entry.file.size, 0);
+
+export const validateAgentEntries = (entries: UploadAgentEntry[]): string | undefined => {
+  if (entries.length === 0) return 'That directory has no uploadable files.';
+
+  if (!entries.some((entry) => entry.path === AGENT_CONFIG_FILENAME)) {
+    return `No ${AGENT_CONFIG_FILENAME} at the top level of that directory.`;
+  }
+
+  if (entries.length > MAX_AGENT_SPEC_FILES) {
+    return `That directory holds ${entries.length} files; the limit is ${MAX_AGENT_SPEC_FILES}. Point at a directory containing only the agent's own files.`;
+  }
+
+  const bytes = totalEntryBytes(entries);
+  if (bytes > MAX_AGENT_SPEC_BYTES) {
+    return `That directory is ${Math.round(bytes / 1000)} KB; the limit is ${Math.round(MAX_AGENT_SPEC_BYTES / 1000)} KB. Point at a directory containing only the agent's own files.`;
+  }
+
+  return undefined;
+};
+
+// Container deployments read every staged file as text and fail on a decode error.
+export const findNonUtf8Path = async (entries: UploadAgentEntry[]): Promise<string | undefined> => {
+  const decoder = new TextDecoder('utf-8', { fatal: true });
+
+  const offenders = await Promise.all(
+    entries.map(async (entry) => {
+      if (entry.path.split('/').pop() === AGENT_SPEC_FILENAME) return undefined;
+      try {
+        decoder.decode(await entry.file.arrayBuffer());
+        return undefined;
+      } catch {
+        return entry.path;
+      }
+    })
+  );
+
+  return offenders.find((path) => path !== undefined);
+};
+
+export class AgentConfigParseError extends Error {}
+
+export const parseAgentConfig = (text: string): Record<string, unknown> => {
+  let parsed: unknown;
+  try {
+    parsed = YAML.parse(text);
+  } catch (error) {
+    throw new AgentConfigParseError(
+      `${AGENT_CONFIG_FILENAME} is not valid YAML: ${(error as Error).message}`
+    );
+  }
+
+  if (typeof parsed !== 'object' || parsed === null || Array.isArray(parsed)) {
+    throw new AgentConfigParseError(`${AGENT_CONFIG_FILENAME} must contain a YAML mapping.`);
+  }
+
+  const config = parsed as Record<string, unknown>;
+  const configFormat = config.config_format;
+  if (configFormat !== FABRIC_CONFIG_FORMAT) {
+    throw new AgentConfigParseError(
+      `${AGENT_CONFIG_FILENAME} must set config_format: ${FABRIC_CONFIG_FORMAT}${
+        typeof configFormat === 'string' ? ` (found ${configFormat})` : ''
+      }.`
+    );
+  }
+
+  return config;
+};
+
+export const agentNameFromConfig = (config: Record<string, unknown>): string | undefined =>
+  typeof config.name === 'string' && config.name.trim() ? config.name.trim() : undefined;
